Add files via upload
This project showcases a portfolio website for Aakash Patel, a passionate web developer. Built with HTML, CSS, and JavaScript, the portfolio highlights Aakash's skills, work experience, and personal details. The website is designed to be responsive and user-friendly, providing an overview of Aakash's professional journey and allowing visitors to contact him easily.